refactor: reorganize script internals into lib package
