+++ scripts/lib/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: MPL-2.0
