+++ scripts/lib/common.py
