+++ scripts/lib/dmg_settings.py
