+++ scripts/tests/test_patch_wine_runtime.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).parents[1]))
 
-import patch_wine_runtime
+from lib import patch_wine_runtime
 
 
 class WineRuntimePatchTests(unittest.TestCase):
